requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file=requirements.txt requirements.in
#
alembic-autogenerate-enums==0.0.2  # via -r requirements.in
alembic==1.4.3            # via flask-migrate
aniso8601==8.0.0          # via flask-restful
appdirs==1.4.4            # via virtualenv
arrow==0.17.0             # via -r requirements.in
asyncpool==1.0            # via -r requirements.in
bcrypt==3.2.0             # via flask-bcrypt
blinker==1.4              # via flask-principal, raven
certifi==2020.11.8        # via requests
cffi==1.14.4              # via bcrypt, cryptography
cfgv==3.2.0               # via pre-commit
chardet==3.0.4            # via requests
click==7.1.2              # via flask
cryptography==3.2.1       # via -r requirements.in, pyopenssl
distlib==0.3.1            # via virtualenv
filelock==3.0.12          # via virtualenv
flake8==3.8.4             # via -r requirements.in
flask-bcrypt==0.7.1       # via -r requirements.in
flask-cors==3.0.9         # via -r requirements.in
flask-migrate==2.5.3      # via -r requirements.in
flask-principal==0.4.0    # via -r requirements.in
flask-replicated==1.4     # via -r requirements.in
flask-restful==0.3.8      # via -r requirements.in
flask-script==2.0.6       # via -r requirements.in
flask-sqlalchemy==2.4.4   # via -r requirements.in, flask-migrate
flask==1.1.2              # via -r requirements.in, flask-bcrypt, flask-cors, flask-migrate, flask-principal, flask-restful, flask-script, flask-sqlalchemy, raven
future==0.18.2            # via -r requirements.in
gunicorn==20.0.4          # via -r requirements.in
identify==1.5.10          # via pre-commit
idna==2.10                # via requests
inflection==0.5.1         # via -r requirements.in
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via -r requirements.in, flask
lockfile==0.12.2          # via -r requirements.in
logmatic-python==0.1.7    # via -r requirements.in
mako==1.1.3               # via alembic
markupsafe==1.1.1         # via jinja2, mako
marshmallow-sqlalchemy==0.24.1  # via -r requirements.in
marshmallow==3.9.1        # via -r requirements.in, marshmallow-sqlalchemy
mccabe==0.6.1             # via flake8
ndg-httpsclient==0.5.1    # via -r requirements.in
nodeenv==1.5.0            # via pre-commit
pre-commit==2.9.1         # via -r requirements.in
psycopg2==2.8.6           # via -r requirements.in
pyasn1-modules==0.2.8     # via python-ldap
pyasn1==0.4.8             # via ndg-httpsclient, pyasn1-modules, python-ldap
pycodestyle==2.6.0        # via flake8
pycparser==2.20           # via cffi
pyflakes==2.2.0           # via flake8
pyjwt==1.7.1              # via -r requirements.in
pyopenssl==19.1.0         # via ndg-httpsclient
python-dateutil==2.8.1    # via alembic, arrow
python-editor==1.0.4      # via alembic
python-json-logger==2.0.1  # via logmatic-python
python-ldap==3.3.1        # via -r requirements.in
pytz==2020.4              # via flask-restful
pyyaml==5.3.1             # via -r requirements.in, pre-commit
raven[flask]==6.10.0      # via -r requirements.in
requests==2.25.0          # via -r requirements.in
retrying==1.3.3           # via -r requirements.in
six==1.15.0               # via -r requirements.in, bcrypt, cryptography, flask-cors, flask-restful, pyopenssl, python-dateutil, retrying, sqlalchemy-utils, virtualenv
sqlalchemy-utils==0.36.8  # via -r requirements.in
sqlalchemy==1.3.20        # via -r requirements.in, alembic, flask-sqlalchemy, marshmallow-sqlalchemy, sqlalchemy-utils
tabulate==0.8.7           # via -r requirements.in
toml==0.10.2              # via pre-commit
urllib3==1.26.2           # via requests
virtualenv==20.2.1        # via pre-commit
werkzeug==1.0.1           # via flask
xmltodict==0.12.0         # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
